Assistant checkpoint: Added person pages and updated game logic with links
Assistant generated file changes:
- game.js: Update INITIAL_GROUPS with page URLs
- james.html: Create James's page
- cyndi.html: Create Cyndi's page
- emily.html: Create Emily's page
- andy.html: Create Andy's page

---

User prompt:

OK, this is all working well. Now we need to add four additional web pages. Each page will be associated with one of the four groups of words. The words data structure should include the url to the page for each group of words. Each page should have an overall title, which will be James, Cyndi, Emily, or Andy. Each page will also contain four photographs. Each photograph will have a title (which will be one of the four words in the group) and a paragraph explaining why that word was chosen.
Finally, when a group of four words is successfully chosen, the description should include a hyperlink to the associated page, so that people will be able to see the pictures.

Can you create this for me please.